LayoutTests/inspector/css/modify-css-property-race-expected.txt

Testing that changes to "style" attribute made from page's JavaScript are ignored when there's a pending change made via Web Inspector.


== Running test suite: ModifyCSSProperty
-- Running test case: ModifyCSSPropertyRace.ChangeInlineStyle
PASS: Height should have increased from 10px.
PASS: Height should have increased from 11px.
PASS: Height should be 100px or more.
PASS: Height should be 101px or more.
